% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_consensus.R
\name{calculate_consensus}
\alias{calculate_consensus}
\title{Calculate the median consensus profile of every subclone}
\usage{
calculate_consensus(df, clusters, mc.cores = 30)
}
\arguments{
\item{df}{The ordered population segmentation profile}

\item{clusters}{The cluster assignment of each single-cell}

\item{mc.cores}{Number of threads.}
}
\value{

}
\description{
Calculate the median consensus profile of every subclone
}
