% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_genomic_classes.R
\name{consensus_genomic_classes}
\alias{consensus_genomic_classes}
\title{For every bin, returns the inferred genomic classification class (cCNA, sCNA, uCNA)}
\usage{
consensus_genomic_classes(consensus_df, ploidy_VAL)
}
\arguments{
\item{consensus_df}{The data frame with the consensus profile for each superclone/subclone}

\item{ploidy_VAL}{The inferred ploidy value for the tumor}
}
\value{
A vector where every bin is classified in cCNA, sCNA or uCNA
}
\description{
For every bin, returns the inferred genomic classification class (cCNA, sCNA, uCNA)
}
\examples{

}
