% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_dataset.R
\name{order_dataset}
\alias{order_dataset}
\title{Orderes the segmentation dataset for downstream usage}
\usage{
order_dataset(popseg_long, clustering, mc.cores = 10)
}
\arguments{
\item{popseg_long}{The resulting data from segmentation}

\item{clustering}{Resulting data frame from}

\item{mc.cores}{Number of threads to be used.}
}
\value{
A list with 2 objects. 1. The ordered segmentation dataset and the ordered clustering dataset
}
\description{
Orderes the segmentation dataset for downstream usage
}
