% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_consensus_heatmap.R
\name{plot_consensus_heatmap}
\alias{plot_consensus_heatmap}
\title{Plot the consensus heatmap with the inferred MRCA profile and gene annotation}
\usage{
plot_consensus_heatmap(
  df,
  clusters,
  ploidy_VAL,
  ploidy_trunc,
  keep_gene = NULL,
  tree_order = NULL,
  plot_title = NULL,
  genomic_classes = NULL
)
}
\arguments{
\item{df}{The data frame with the inferred consensus profiles}

\item{clusters}{data.frame. Resulting from clustering}

\item{ploidy_VAL}{Numeric. The inferred ploidy value}

\item{ploidy_trunc}{Numeric. Value that will truncate levels higer than it for visualization}

\item{keep_gene}{Genes that will be kept on the annotation. Must be provided otherwise it will annotate all genes in hg19}

\item{tree_order}{Ordering of subclones resulting from the me tree. Will be used to order the heatmap}

\item{plot_title}{Title of the plot}

\item{genomic_classes}{Character. Vector with genomic regions classification}
}
\value{

}
\description{
Plot the consensus heatmap with the inferred MRCA profile and gene annotation
}
