% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot the single-cell heatmap}
\usage{
plot_heatmap(
  df,
  ploidy_VAL,
  ploidy_trunc,
  clusters,
  genomic_classes = NULL,
  keep_gene,
  tree_order = NULL,
  show_legend = FALSE
)
}
\arguments{
\item{df}{The ordered segmented single-cell data frame}

\item{ploidy_VAL}{The inferred sample ploidy value}

\item{ploidy_trunc}{Numeric. Value that will truncate levels higer than it for visualization}

\item{clusters}{Data frame with clustering information}

\item{genomic_classes}{Vector with genomic regions classification}

\item{keep_gene}{Genes that will be kept on the annotation. Must be provided otherwise it will annotate all genes in hg19}

\item{tree_order}{Ordering of subclones resulting from the me tree. Will be used to order the heatmap}

\item{show_legend}{Boolean. Display heatmap legend}
}
\value{
A ComplexHeatmap with the single-cells annotated by clustering.
}
\description{
Plot the single-cell heatmap
}
