% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_umap.R
\name{plot_umap}
\alias{plot_umap}
\title{Plots the resulting umap and colors with the clustering information}
\usage{
plot_umap(umap_df, clustering)
}
\arguments{
\item{clustering}{}
}
\value{
A ggplot object with a umap scatterplot.
}
\description{
Plots the resulting umap and colors with the clustering information
}
