% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_clustering.R
\name{run_clustering}
\alias{run_clustering}
\title{Run HDBSCAN clustering on the resulting embedding from umap}
\usage{
run_clustering(umap_df, k_snn_major = 35, k_snn_minor = 17)
}
\arguments{
\item{umap_df}{Data frame with the resulting embedding from \code{run_umap}.}

\item{k_snn_major}{nearest neighbors K value. Passed to build an SNN graph that will be used to determine superclones}

\item{k_snn_minor}{nearest neighbors K value. Passed to \code{dbscan::hdbscan} minPts argument.}
}
\value{
A data frame containing the superclone and subclone assignment for each cell
}
\description{
Run HDBSCAN clustering on the resulting embedding from umap
}
\examples{


}
