% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_me_tree.R
\name{run_me_tree}
\alias{run_me_tree}
\title{Runs a minimal evolution tree algorithm for the consensus data frame}
\usage{
run_me_tree(consensus_df, clusters, ploidy_VAL, rotate_nodes = NULL)
}
\arguments{
\item{consensus_df}{The consensus data frame from the segmented dataset}

\item{clusters}{The results from clustering}

\item{ploidy_VAL}{The inferred ploidy value. Will be used to scale the data}

\item{rotate_nodes}{Nodes to be rotated}
}
\value{

}
\description{
Runs a minimal evolution tree algorithm for the consensus data frame
}
