% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_umap.R
\name{run_umap}
\alias{run_umap}
\title{Run Umap dimension reduction}
\usage{
run_umap(
  popseg_long,
  ploidy_VAL = NULL,
  umap_dist = "manhattan",
  umap_min_dist = 0,
  umap_spread = 1,
  umap_n_neighbors = 40,
  mc.cores = 10,
  seed = 55,
  round = FALSE
)
}
\arguments{
\item{popseg_long}{The resulting data from segmentation}

\item{ploidy_VAL}{Optional ploidy value to scale the segment ratios values by its ploidy}

\item{umap_dist}{Distance metric used by umap. Defaults to manhattan}

\item{umap_min_dist}{UMAP min dist parameter}

\item{umap_spread}{UMAP spread parameter}

\item{umap_n_neighbors}{UMAP n_neighbors parameter}

\item{mc.cores}{Number of threads to be used}

\item{seed}{Seed, defaults to 55}

\item{round}{Optional in case data is ploidy scaled values are rounded to the nearest integer}
}
\value{
A data frame with the resulting embedding from umap and a column containing the cell names.
}
\description{
Run Umap dimension reduction
}
\examples{

}
